Track per block position and color
Now we have changed the model to be tracking
per cube block's position and color. This makes
sure we can always traversal back along cosets.

The new problem now is, the generators grow
exponentially along with we calculate through
the stablizer chain. This is the general problem
of naively applying Schreier subgroup lemma.