Employ Incremental Schreier-Sims algorithm
Previous problem is, naively applying Schreier subgroup
lemma, results in that generator count grows exponentially
along the stablizer chain.

In this patch, we applied the Incremental Schreier-Sims
algorithm, so that the redundant generators will be
detected and filtered out in early stage.

New problems are, the overall calculation is still slow,
because after several steps along the stablizer chain,
the generators grow longer, and then the speed of
calculation slows down.