Reduce generator length by matching cache
Generator length too long is the main reason to drag
down the calculation speed at deeper GSteps.

We built a matching cache of short cube actions, and
for any long cube action, we discover any section of
it can be replaced to shorter cube actions.

This method is not very effective. Despite the very
high computation complexity, it only reduce 10% or
less length of each generator. (The good side is the
10% propagates to each GStep deeper.)