Use ActionMap to accelerate long CubeAction.Act
The problem is, as we proceed along the stablizer chain,
generators are getting longer and longer. Majority time
was spent doing such long CubeAction's Act method.

However, no matter how long a generator is, the CubeAction
is essentially a permutation mapping, which can be applied
in constant time.

We have constructed the algorithm to obtain the permutation
mapping from a CubeAction, to apply the permuationtes after
transform, and to obtain the new Block states after transform.

The calculation speed has been increased exponentially for
long generators acting.