Adjusting init generator sequence, etc
As observed, when 3 of op F, U, L are added,
there is an explosion of discovered new cosets
and new generators.

So we adjust the sequence of init generators,
to add all of F, U, L in early stage, and to
add them balancedly. This should help we reach
wider cube states with shorter generator length.

Besides, we added a new GroupTest case to probe
which is the proper config of
OpCountForAccelerationMap. This should improve
speed for calculating short generators that are
inappropriate to use AccelerationMap.