Simplify CubeAction Reverse Operations
Reverse operation is the main reason to bloat
the length of CubeAction. Along the tree of
CubeAction being composed, the Reverse operation
can be sunk down along the tree, and many can
be neutralized.

But the reality doesn't work out as well as
the theory. In an observed case, the tree is
100+ layers deep. Many nodes are shared.

Sinking Reverse operation down involves two many
"new" because all CubeActions along the path
need to be reversed one by one.

We also tried using FlipReversed to lazy flag
an CubeAction to be reversed rather than "new"
always. However, since many tree nodes are actually
shared, flags interfere with each other.

Naviely deep copy entire tree is also impossible,
because the 100+ layer deep tree is too large.