GetOpsReversed for online SimplifyNoops
Previous bottleneck is Enumerable.Reverse(GetOps()),
where C# tries to buffer all middle results before
Reverse. This is infeasible for very large CubeAction.

We implement GetOpsReversed, an reverse iterator
for CubeAction. Then in SimplifyNoops we can online
ingest and dedup operations.

This is not enough to SimplifyNoops very large
CubeActions, but now we are at least able to print
each operations in it online.