Print large CubeAction with formula
Though we don't have effective way to
print the 100+ layer deep CubeAction
by each operation, we can print it by
formula.

It proves effective to print as formula.
This reveals the internal structure of
various coset actions and steps to solve
the Rubik's Cube.

Now we can consider the work is done.